+++ src/test/converter2/issues/gh2811.ts
@@ -1,0 +1,6 @@
+export const abc = 123;
+
+export { abc as rename1 };
+
+import { rename1 } from "./gh2811.js";
+export { rename1 as rename2 };