+++ CHANGELOG.md
@@ -7,6 +7,7 @@ title: Changelog
 ### Bug Fixes
 
 -   `@include` and `@includeCode` now work in the readme file, #2814.
+-   TypeDoc will now avoid making references to references, #2811.
 
 ## v0.27.5 (2024-12-14)